Move SceneRenderer to engine, update dependency graph
