+++ lib/engine/CMakeLists.txt
@@ -35,12 +35,14 @@ endif()
 
 target_sources(${PROJECT_NAME} PRIVATE
   include/${target_prefix}/engine/engine.hpp
+  include/${target_prefix}/engine/scene_renderer.hpp
 
   include/${target_prefix}/engine/editor/common.hpp
   include/${target_prefix}/engine/editor/inspector.hpp
   include/${target_prefix}/engine/editor/log.hpp
   include/${target_prefix}/engine/editor/scene_tree.hpp
 
+  src/scene_renderer.cpp
   src/engine.cpp
 
   src/editor/common.cpp
