+++ lib/scene/CMakeLists.txt
@@ -19,7 +19,7 @@ target_include_directories(${PROJECT_NAME}
 
 target_link_libraries(${PROJECT_NAME}
   PUBLIC
-  ${target_prefix}::render
+  ${target_prefix}::vk
   djson::djson
 
   PRIVATE
@@ -41,7 +41,6 @@ target_sources(${PROJECT_NAME} PRIVATE
   include/${target_prefix}/scene/node_data.hpp
   include/${target_prefix}/scene/node.hpp
   include/${target_prefix}/scene/rect.hpp
-  include/${target_prefix}/scene/scene_renderer.hpp
   include/${target_prefix}/scene/scene.hpp
   include/${target_prefix}/scene/transform.hpp
 
@@ -50,7 +49,6 @@ target_sources(${PROJECT_NAME} PRIVATE
 
   src/camera.cpp
   src/material.cpp
-  src/scene_renderer.cpp
   src/scene.cpp
   src/transform.cpp
 )
